16 jan

+++ TuristRegistar/Data/Models/OccupancyBasedPrices.cs
@@ -9,7 +9,7 @@ namespace TuristRegistar.Data.Models
     {
         public int Id { get; set; }
         public int Occupancy { get; set; }
-        public float PricePerNight { get; set; }
+        public Decimal PricePerNight { get; set; }
 
         public int OccunapncyBasedPricingId { get; set; }
         public OccupancyBasedPricing OccunapncyBasedPricing { get; set; }
